ag71xx: specify new parameter frag_size parameter for build_skb()
Commit d3836f21b0af5513ef55701dd3f50b8c42e44c7a allows "skb->head to be
a page fragment", and the frag_size parameter to be 0.

"build_skb() is changed to accept a frag_size argument. Drivers willing
to provide a page fragment instead of kmalloc() data will set a non zero
value, set to the fragment size."

Set a value of 0 for frag_size indicating kmalloc() data.

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>