emgd: Fixes for 3.10 kernel
Various fixes to get emgd-1.18 compiling with the 3.10 kernel. The
main changes were to replace the defunct create_proc_read_entry() etc
with proc_create_data() and PDE_DATA() and fix up proc entry removal
in light of the fact that proc_dir_entry is now opaque and used to
contain the name.

The proc files we need to worry about are handled by the above changes
- there are a couple sections of proc removal and write_proc() code
that were removed because they're only used by DEBUG code and we don't
care about fixing up debug code.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>