emgd-1.18: Fix ioctl handling
commit e4fda9f [drm: Perform ioctl command validation on the stored
kernel values] causes a couple problems that need to be fixed.

One is that ioctl.cmd_drv needs to be given an appropriate value
instead of 0 since it's now used as the cmd value in the handler -
EMGD_IOCTL_DEF_DRV() is introduced and used instead of
EMGD_IOCTL_DEF().

The other problem is that the above exposes the fact that the cmds
that userspace is sending are not what the kernel says it's expecting
e.g. the kernel says IOR but what's being passed in is IOWR, which
causes different (correct) kernel handling, but that doesn't match the
kernel code.  So the kernel code is wrong, and if not corrected makes
video inoperable.  Since we don't have access to the userland code
that sends the commands, we need to change the affected kernel ioctl
values to match.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>