serial: 8250_dw: mask UART TX completion intr in byt_set_termios
The code masking for TX completion interrupt was added by commit
"06d8641 ACPI / LPSS: mask the UART TX completion interrupt".
At that time, this code executes during initialization. Regression
test on power management suspend/resume reveals the need for this
bit to be masked again during controller resume. So, instead of
doing the mask during Baytrail LPSS initialization, put it in
byt_set_termios in the 8250 platform device code to make sure that
the bit is properly set even on the event of system suspend/resume.

Signed-off-by: Maurice Petallo <mauricex.r.petallo@intel.com>
Change-Id: Ie262e3a69c1a4ea582a7a7206f1c36745fe75618
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/3950
Reviewed-by: Kweh, Hock Leong <hock.leong.kweh@intel.com>
Reviewed-by: Wan Mohamad, Wan Ahmad Zainie <wan.ahmad.zainie.wan.mohamad@intel.com>