dmaengine: dw: move clock operations to platform.c
On BayTrail platform DMA is not functional in the PCI mode, whereby it always
failed and exit at the point when it tries to get a clock. It causes the PCI
mode probe to exit with the error message:
	dw_dmac_pci: probe of 0000:00:1e.0 failed with error -2

This patch moves clock operations to where it belongs to. Thus, the clock is
provided only in ACPI / non-PCI cases.

Reported-by: Chew, Chiau Ee <chiau.ee.chew@intel.com>
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
(cherry picked from commit a15636e83eb0dedefcb1221be729023e4c281748)

Conflicts:

	drivers/dma/dw/core.c
	- Backport to Kernel 3.10.61-ltsi. Make the patch apply but
	  retaining the essence of it which is to remove clk related
	  stuff out of core.c file.

Signed-off-by: Maurice Petallo <mauricex.r.petallo@intel.com>
Change-Id: Ia696de420231608cd820fd1e227191db71540a68
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/4363
Reviewed-by: Wan Mohamad, Wan Ahmad Zainie <wan.ahmad.zainie.wan.mohamad@intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>